Revert "Refactor sink task integration tests to be more readable (#543)"
This reverts commit 1b24187c864322549e60963db58bc75bb8ca3924.
